# Conflicts:
#	src/main/java/com/thefitnation/service/UserDemographicService.java

+++ src/main/java/com/thefitnation/service/WorkoutTemplateService.java
@@ -1,27 +1,16 @@
 package com.thefitnation.service;
 
-import com.thefitnation.domain.User;
-import com.thefitnation.domain.UserDemographic;
-import com.thefitnation.domain.UserWorkoutTemplate;
-import com.thefitnation.domain.WorkoutTemplate;
-import com.thefitnation.repository.UserDemographicRepository;
-import com.thefitnation.repository.UserRepository;
-import com.thefitnation.repository.UserWorkoutTemplateRepository;
-import com.thefitnation.repository.WorkoutTemplateRepository;
-import com.thefitnation.security.SecurityUtils;
-import com.thefitnation.service.dto.WorkoutTemplateDTO;
-import com.thefitnation.service.dto.WorkoutTemplateWithChildrenDTO;
-import com.thefitnation.service.mapper.WorkoutTemplateMapper;
-import com.thefitnation.service.mapper.WorkoutTemplateWithChildrenMapper;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.stereotype.Service;
-
-import javax.swing.text.html.Option;
-import java.util.Optional;
+import com.thefitnation.domain.*;
+import com.thefitnation.repository.*;
+import com.thefitnation.security.*;
+import com.thefitnation.service.dto.*;
+import com.thefitnation.service.mapper.*;
+import java.time.*;
+import java.util.*;
+import org.slf4j.*;
+import org.springframework.data.domain.*;
+import org.springframework.stereotype.*;
+import org.springframework.transaction.annotation.*;
 
 /**
  * Service Implementation for managing WorkoutTemplate.
@@ -33,17 +22,21 @@ public class WorkoutTemplateService {
     private final Logger log = LoggerFactory.getLogger(WorkoutTemplateService.class);
 
     private final UserRepository userRepository;
-
     private final UserDemographicRepository userDemographicRepository;
-
     private final WorkoutTemplateRepository workoutTemplateRepository;
-
     private final UserWorkoutTemplateRepository userWorkoutTemplateRepository;
-
     private final WorkoutTemplateMapper workoutTemplateMapper;
-
     private final WorkoutTemplateWithChildrenMapper workoutTemplateWithChildrenMapper;
 
+    /**
+     * Constructor
+     * @param userRepository for getting user data
+     * @param userDemographicRepository for getting UserDemographic data
+     * @param workoutTemplateRepository for getting WorkoutTemplate data
+     * @param userWorkoutTemplateRepository for getting User data
+     * @param workoutTemplateMapper to map WorkoutTemplate to and from WorkoutTemplateDTO
+     * @param workoutTemplateWithChildrenMapper to map WorkoutTemplateWithChildren to and from WorkoutTemplateWithChildrenDTO
+     */
     public WorkoutTemplateService(UserRepository userRepository, UserDemographicRepository userDemographicRepository, WorkoutTemplateRepository workoutTemplateRepository, UserWorkoutTemplateRepository userWorkoutTemplateRepository, WorkoutTemplateMapper workoutTemplateMapper, WorkoutTemplateWithChildrenMapper workoutTemplateWithChildrenMapper) {
         this.userRepository = userRepository;
         this.userDemographicRepository = userDemographicRepository;
@@ -61,18 +54,38 @@ public class WorkoutTemplateService {
      */
     public WorkoutTemplateDTO save(WorkoutTemplateDTO workoutTemplateDTO) {
         log.debug("Request to save WorkoutTemplate : {}", workoutTemplateDTO);
+
+        WorkoutTemplate workoutTemplate = workoutTemplateMapper.workoutTemplateDTOToWorkoutTemplate(workoutTemplateDTO);
+
+        workoutTemplate.setUserDemographic(userDemographicRepository.findOneByLogin(SecurityUtils.getCurrentUserLogin()));
+        workoutTemplate.setCreatedOn(LocalDate.now());
+        workoutTemplate.setLastUpdated(LocalDate.now());
+
+        workoutTemplate = workoutTemplateRepository.save(workoutTemplate);
+        return workoutTemplateMapper.workoutTemplateToWorkoutTemplateDTO(workoutTemplate);
+    }
+
+    /**
+     * Update a workoutTemplate.
+     *
+     * @param workoutTemplateDTO the entity to save
+     * @return the persisted entity
+     */
+    public WorkoutTemplateDTO update(WorkoutTemplateDTO workoutTemplateDTO) {
+        log.debug("Request to update WorkoutTemplate : {}", workoutTemplateDTO);
         if (workoutTemplateDTO.getUserDemographicId() == null) {
             Optional<User> user = userRepository.findOneByLogin(SecurityUtils.getCurrentUserLogin());
+
             if (user.isPresent()) {
                 UserDemographic userDemographic = userDemographicRepository.findOneByUserWithEagerRelationships(user.get().getId());
                 workoutTemplateDTO.setUserDemographicId(userDemographic.getId());
             }
         }
-        WorkoutTemplate workoutTemplate = workoutTemplateMapper.workoutTemplateDTOToWorkoutTemplate(workoutTemplateDTO);
 
+        WorkoutTemplate workoutTemplate = workoutTemplateMapper.workoutTemplateDTOToWorkoutTemplate(workoutTemplateDTO);
+        workoutTemplate.setLastUpdated(LocalDate.now());
         workoutTemplate = workoutTemplateRepository.save(workoutTemplate);
-        WorkoutTemplateDTO result = workoutTemplateMapper.workoutTemplateToWorkoutTemplateDTO(workoutTemplate);
-        return result;
+        return workoutTemplateMapper.workoutTemplateToWorkoutTemplateDTO(workoutTemplate);
     }
 
     /**
@@ -89,6 +102,19 @@ public class WorkoutTemplateService {
     }
 
     /**
+     * find All by Login
+     * @param pageable options for returning WorkoutTemplateDTOs
+     * @return a Page of WorkoutTemplateDTOs
+     */
+    public Page<WorkoutTemplateDTO> findAllByLogin(Pageable pageable) {
+        log.debug("Request to get all WorkoutTemplates by current logged in user.");
+        String login = SecurityUtils.getCurrentUserLogin();
+        Page<WorkoutTemplate> result = workoutTemplateRepository.findAllByCurrentLoggedInUser(login, pageable);
+        return result.map(workoutTemplateMapper::workoutTemplateToWorkoutTemplateDTO);
+
+    }
+
+    /**
      *  Get one workoutTemplate by id.
      *
      *  @param id the id of the entity
@@ -97,7 +123,8 @@ public class WorkoutTemplateService {
     @Transactional(readOnly = true)
     public WorkoutTemplateWithChildrenDTO findOne(Long id) {
         log.debug("Request to get WorkoutTemplate : {}", id);
-        WorkoutTemplate workoutTemplate = workoutTemplateRepository.findOne(id);
+        String login = SecurityUtils.getCurrentUserLogin();
+        WorkoutTemplate workoutTemplate = workoutTemplateRepository.findOne(login, id);
         WorkoutTemplateWithChildrenDTO workoutTemplateDTO = workoutTemplateWithChildrenMapper.workoutTemplateToWorkoutTemplateWithChildrenDTO(workoutTemplate);
         return workoutTemplateDTO;
     }
@@ -109,8 +136,9 @@ public class WorkoutTemplateService {
      */
     public void delete(Long id) {
         log.debug("Request to delete WorkoutTemplate : {}", id);
-        removeWorkoutTemplateFromRelatedItems(id);
-        workoutTemplateRepository.delete(id);
+
+        if (workoutTemplateRepository.findOne(SecurityUtils.getCurrentUserLogin(), id).getId() != null)
+            workoutTemplateRepository.delete(id);
     }
 
     public void removeWorkoutTemplateFromRelatedItems(Long id) {
